requirements.txt
google-adk==1.30.0
yfinance